backend/requirements.txt
gunicorn==21.2.0
sorl-thumbnail==12.10.0
Django==3.2.3
django-filter==23.3
djangorestframework==3.14.0
djoser==2.2.2
drf_extra_fields==3.7.0
python-dotenv==1.0.0
psycopg2-binary==2.8.6
Pillow==10.1.0
flake8==6.0.0